+++ lib/compression/test/test_compression_lzma.cpp
@@ -12,7 +12,7 @@ int main(int argc, char *argv[])
   VSX_UNUSED(argc);
   VSX_UNUSED(argv);
 
-  srand(time(0x0));
+  srand((unsigned int)time(0x0));
 
   vsx_ma_vector<unsigned char> data;
 
@@ -36,7 +36,7 @@ int main(int argc, char *argv[])
 
   foreach(data, i)
     if (data[i] != uncompressed[i])
-      vsx_printf(L"data differs on index %d\n", i);
+      vsx_printf(L"data differs on index %zu\n", i);
 
   return 0;
 }
